Feat: Change vCPU to be float64 not int
Some VMs have float values for instance vCPU count if they
have shared vCPUs

+++ pkg/types/v2/types.go
@@ -7,7 +7,7 @@ type Wattage struct {
 
 type Instance struct {
 	Kind                 string
-	VCPU                 int
+	VCPU                 float64
 	MemoryGB             float64
 	GPUCount             int
 	GPUMemoryGB          float64
